test: handlerUtils 로 분리
